+++ examples/nextjs-page-router-example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # nextjs-page-router-example
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - meta-cloud-api@3.0.0
+
 ## 0.1.7
 
 ### Patch Changes